Refactor provider utils
