+++ packages/checkout/widgets-lib/src/widgets/wallet/views/Settings.tsx
@@ -2,7 +2,7 @@ import { Box, Button } from '@biom3/react';
 import { useContext, useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useWalletConnect } from 'lib/hooks/useWalletConnect';
-import { isWalletConnectProvider } from 'lib/providerUtils';
+import { isWalletConnectProvider } from 'lib/provider';
 import { FooterLogo } from '../../../components/Footer/FooterLogo';
 import { HeaderNavigation } from '../../../components/Header/HeaderNavigation';
 import { SimpleLayout } from '../../../components/SimpleLayout/SimpleLayout';
